Separate out miral-server-libs
