+++ examples/miral-kiosk/CMakeLists.txt
@@ -1,9 +1,3 @@
-include_directories(
-    ../miral-shell
-    ${PROJECT_SOURCE_DIR}/include/miral
-    ${PROJECT_SOURCE_DIR}/include/client
-)
-
 mir_add_wrapped_executable(miral-kiosk
     kiosk_main.cpp
     kiosk_window_manager.cpp    kiosk_window_manager.h
