+++ examples/miral-server-lib/decoration_provider.cpp
