+++ examples/miral-server-lib/decoration_provider.h
