+++ examples/miral-server-lib/floating_window_manager.cpp
