+++ examples/miral-server-lib/mir-peace-wall.png
