+++ examples/miral-server-lib/sw_splash.cpp
