+++ examples/miral-server-lib/sw_splash.h
