+++ examples/miral-server-lib/titlebar_config.h
