+++ examples/miral-shell/CMakeLists.txt
@@ -1,10 +1,5 @@
 find_package(PIL REQUIRED)
 
-include_directories(
-        ${PROJECT_SOURCE_DIR}/include/miral
-        ${PROJECT_SOURCE_DIR}/include/client
-)
-
 function(png2header png header varname)
     add_custom_command(
         OUTPUT ${header}
@@ -14,12 +9,6 @@ function(png2header png header varname)
 endfunction()
 
 add_subdirectory(desktop)
-
-png2header(
-    ${CMAKE_CURRENT_SOURCE_DIR}/mir-peace-wall.png
-    ${CMAKE_CURRENT_SOURCE_DIR}/mir_peace_wall.h
-    mir_peace_wall
-)
 
 add_custom_target(miral-run ALL
     cp ${CMAKE_CURRENT_SOURCE_DIR}/miral-run.sh ${CMAKE_RUNTIME_OUTPUT_DIRECTORY}/miral-run
@@ -60,23 +49,6 @@ add_custom_target(miral-app ALL
 install(PROGRAMS ${CMAKE_RUNTIME_OUTPUT_DIRECTORY}/miral-app
     DESTINATION ${CMAKE_INSTALL_PREFIX}/bin
 )
-
-add_library(miral-shell-lib STATIC
-    tiling_window_manager.cpp   tiling_window_manager.h
-    floating_window_manager.cpp floating_window_manager.h
-    decoration_provider.cpp     decoration_provider.h
-    titlebar_config.cpp         titlebar_config.h
-    sw_splash.cpp               sw_splash.h
-    mir_peace_wall.h
-)
-
-set_source_files_properties(decoration_provider.cpp PROPERTIES COMPILE_FLAGS
-    "${CMAKE_CXX_FLAGS}  -Wno-overlength-strings")
-
-pkg_check_modules(FREETYPE freetype2 REQUIRED)
-target_include_directories(miral-shell-lib PRIVATE ${FREETYPE_INCLUDE_DIRS})
-target_compile_definitions(miral-shell-lib PRIVATE -DTYPO_SUPPORTS_FREETYPE)
-target_link_libraries(miral-shell-lib miral ${FREETYPE_LIBRARIES})
 
 mir_add_wrapped_executable(miral-shell
     shell_main.cpp
